+++ src/http-message/src/ContentType.php
@@ -18,7 +18,7 @@ final class ContentType
     public const HTML = 'text/html';
     public const TEXT = 'text/plain';
     public const JSON = 'application/json';
-    public const XML = 'application/xml';
+    public const XML  = 'application/xml';
     public const FORM = 'application/x-www-form-urlencoded';
 
     // Content types mapping