+++ src/stdlib/test/unit/Helper/StringHelperTest.php
@@ -20,4 +20,19 @@ class StringHelperTest extends TestCase
         $name = Str::getClassName('App\Http\Controller\UserController', '');
         $this->assertSame('App\Http\Controller\UserController', $name);
     }
+
+    public function testRmPharPrefix(): void
+    {
+        $path = Str::rmPharPrefix('phar:///path/to/some.phar/vendor/composer');
+        $this->assertSame('/path/to/vendor/composer', $path);
+
+        $path = Str::rmPharPrefix('phar:///path/to/some.phar/vendor/composer/');
+        $this->assertSame('/path/to/vendor/composer/', $path);
+
+        $path = Str::rmPharPrefix('phar:///path/to/some.phar/vendor/composer/.env');
+        $this->assertSame('/path/to/vendor/composer/.env', $path);
+
+        $path = Str::rmPharPrefix('/vendor/composer/.env');
+        $this->assertSame('/vendor/composer/.env', $path);
+    }
 }