+++ src/test/src/Concern/CommonTestAssertTrait.php
@@ -1,0 +1,35 @@
+<?php declare(strict_types=1);
+
+namespace Swoft\Test\Concern;
+
+use PHPUnit\Framework\Assert;
+
+/**
+ * Class CommonTestAssertTrait
+ *
+ * @since 2.0
+ */
+trait CommonTestAssertTrait
+{
+    /**
+     * @param string $haystack
+     * @param string $needle
+     */
+    public function assertContainString(string $haystack, string $needle): void
+    {
+        $exist = \strpos($haystack, $needle) !== false;
+
+        Assert::assertTrue($exist);
+    }
+
+    /**
+     * @param array $haystack
+     * @param mixed $needle
+     */
+    public function assertArrayContainValue(array $haystack, $needle): void
+    {
+        $exist = \in_array($needle, $haystack, true);
+
+        Assert::assertTrue($exist);
+    }
+}