+++ src/websocket-server/src/MessageParser/JsonParser.php
@@ -5,6 +5,7 @@ namespace Swoft\WebSocket\Server\MessageParser;
 use Swoft\Bean\Annotation\Mapping\Bean;
 use Swoft\Stdlib\Helper\JsonHelper;
 use Swoft\WebSocket\Server\Contract\MessageParserInterface;
+use Swoft\WebSocket\Server\Message\Message;
 
 /**
  * Class JsonParser
@@ -14,31 +15,40 @@ use Swoft\WebSocket\Server\Contract\MessageParserInterface;
 class JsonParser implements MessageParserInterface
 {
     /**
-     * @param array $data
+     * @param Message $message
+     *
      * @return string
      */
-    public function encode($data): string
+    public function encode(Message $message): string
     {
-        return JsonHelper::encode($data);
+        return JsonHelper::encode($message->toArray());
     }
 
     /**
      * Decode data to array.
+     *
      * @param string $data Message data. It's {@see \Swoole\WebSocket\Frame::$data)
-     * @return array
+     *
+     * @return Message [
      *                     [
-     *                     'cmd'  => 'home.index', // message command. it's must exists.
-     *                     'data' => message data,
-     *                     ...
+     *                      'cmd'  => 'home.index', // message command. it's must exists.
+     *                      'data' => message data,
+     *                      ...
      *                     ]
+     *                  ]
+     * @throws \ReflectionException
+     * @throws \Swoft\Bean\Exception\ContainerException
      */
-    public function decode(string $data): array
+    public function decode(string $data): Message
     {
+        $cmd = '';
         $map = JsonHelper::decode($data, true);
 
-        return isset($map['cmd']) ? $map : [
-            'cmd'  => '',
-            'data' => $map,
-        ];
+        if (isset($map['cmd'])) {
+            $cmd = (string)$map['cmd'];
+            unset($map['cmd']);
+        }
+
+        return Message::new($cmd, $map['data'] ?? $map);
     }
 }