+++ src/websocket-server/src/Swoole/MessageListener.php
@@ -9,8 +9,10 @@ use Swoft\Server\Swoole\MessageInterface;
 use Swoft\Session\Session;
 use Swoft\SwoftEvent;
 use Swoft\WebSocket\Server\Context\WsMessageContext;
-use Swoft\WebSocket\Server\WsDispatcher;
+use Swoft\WebSocket\Server\Message\Request;
+use Swoft\WebSocket\Server\Message\Response;
 use Swoft\WebSocket\Server\WsErrorDispatcher;
+use Swoft\WebSocket\Server\WsMessageDispatcher;
 use Swoft\WebSocket\Server\WsServerEvent;
 use Swoole\Websocket\Frame;
 use Swoole\Websocket\Server;
@@ -27,6 +29,7 @@ class MessageListener implements MessageInterface
     /**
      * @param Server $server
      * @param Frame  $frame
+     *
      * @throws \Throwable
      */
     public function onMessage(Server $server, Frame $frame): void
@@ -34,23 +37,25 @@ class MessageListener implements MessageInterface
         $fd  = $frame->fd;
         $sid = (string)$fd;
 
+        $req = Request::new($frame);
+        $res = Response::new($fd);
         /** @var WsMessageContext $ctx */
-        $ctx = WsMessageContext::new($frame);
+        $ctx = WsMessageContext::new($req, $res);
 
         // Storage context
         Context::set($ctx);
         // Bind cid => sid(fd)
         Session::bindCo($sid);
 
-        /** @var WsDispatcher $dispatcher */
-        $dispatcher = BeanFactory::getSingleton('wsDispatcher');
+        /** @var WsMessageDispatcher $dispatcher */
+        $dispatcher = BeanFactory::getSingleton('wsMsgDispatcher');
 
         try {
             \server()->log("Message: conn#{$fd} received message: {$frame->data}", [], 'debug');
             \Swoft::trigger(WsServerEvent::MESSAGE_BEFORE, $fd, $server, $frame);
 
             // Parse and dispatch message
-            $dispatcher->message($server, $frame);
+            $dispatcher->dispatch($server, $frame);
 
             \Swoft::trigger(WsServerEvent::MESSAGE_AFTER, $fd, $server, $frame);
         } catch (\Throwable $e) {